refactor: JumpingPiece 관련 패키징
