+++ src/main/java/domain/piece/jumping/Knight.java
@@ -1,7 +1,8 @@
-package domain.piece;
+package domain.piece.jumping;
 
 import domain.Direction;
 import domain.Team;
+import domain.piece.JumpingPiece;
 
 import java.util.List;
